mobile reader: extract useReaderInfiniteScroll
Owns nextChapterRef + appendChapter prefetch + enable/disable WebView
sentinel. Screen calls enableForChapter on 'loaded' and loadNext on
'requestNextChapter'.
